requirements.txt
undetected-chromedriver
beautifulsoup4
argostranslate
streamlit
pandas
plotly
SQLAlchemy
requests
schedule
python-dotenv
fastapi
uvicorn